Rename component files because not use formik
