+++ src/ts/index.tsx
@@ -1,13 +1,13 @@
 import * as React from "react";
 import * as ReactDOM from "react-dom";
 
-import FormikMain from './components/FormikMain';
+import Form from './components/Form';
 
 import '../stylesheet/index.css';
 
 ReactDOM.render(
     <div>
-        <FormikMain query={window.location.search} />
+        <Form query={window.location.search} />
     </div>,
     document.getElementById("content")
 );
